requirements.txt
requests>=2.31.0
pyyaml>=6.0
jinja2>=3.1.0
